+++ app/src/main/java/com/example/charo_android/presentation/ui/charo/follow/CharoListFollowerFragment.kt
@@ -1,4 +1,4 @@
-package com.example.charo_android.presentation.ui.charo
+package com.example.charo_android.presentation.ui.charo.follow
 
 import android.content.Intent
 import android.os.Bundle
@@ -8,13 +8,14 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.activityViewModels
-import com.example.charo_android.R
-import com.example.charo_android.databinding.FragmentCharoListFollowingBinding
+import com.example.charo_android.databinding.FragmentCharoListFollowerBinding
 import com.example.charo_android.hidden.Hidden
+import com.example.charo_android.presentation.ui.charo.viewmodel.CharoViewModel
+import com.example.charo_android.presentation.ui.charo.adapter.CharoListFollowAdapter
 import com.example.charo_android.presentation.ui.main.MainActivity
 
-class CharoListFollowingFragment : Fragment() {
-    private var _binding: FragmentCharoListFollowingBinding? = null
+class CharoListFollowerFragment : Fragment() {
+    private var _binding: FragmentCharoListFollowerBinding? = null
     private val binding get() = _binding!!
     private val charoViewModel: CharoViewModel by activityViewModels()
     private val followAdapter = CharoListFollowAdapter() {
@@ -42,10 +43,10 @@ class CharoListFollowingFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentCharoListFollowingBinding.inflate(layoutInflater, container, false)
+        _binding = FragmentCharoListFollowerBinding.inflate(layoutInflater, container, false)
         initRecyclerView()
         charoViewModel.followData.observe(viewLifecycleOwner, {
-            followAdapter.itemList.addAll(it.following)
+            followAdapter.itemList.addAll(it.follower)
             Log.d("itemList.size", followAdapter.itemList.size.toString())
             followAdapter.notifyDataSetChanged()
         })
@@ -59,7 +60,7 @@ class CharoListFollowingFragment : Fragment() {
 
     private fun initRecyclerView() {
         with(binding) {
-            recyclerViewCharoListFollowing.adapter = followAdapter
+            recyclerViewCharoListFollower.adapter = followAdapter
             followAdapter.notifyDataSetChanged()
         }
     }
